carbon: Refactor layout
