+++ carbon/components/Layout/ChangeLanguageButton/styles.module.scss
@@ -1,0 +1,27 @@
+@use "~@klimadao/lib/theme/breakpoints.scss";
+@use "theme/typography.module.scss";
+
+.topBarTooltip {
+  @include typography.body1;
+  text-align: left;
+  background: var(--brand-white);
+  border-radius: 0.8rem;
+  padding: 0 1.2rem;
+  min-width: 15rem;
+  box-shadow:
+    0 0.4rem 0.4rem rgba(0, 0, 0, 0.24),
+    0 0 0.4rem rgba(0, 0, 0, 0.12);
+
+  @media (max-width: 32.5rem) {
+    &.tippy-box {
+      max-width: 24rem !important; // fix x-overflow bug on small mobile
+    }
+  }
+
+  div[aria-describedby="tooltip-content"] {
+    padding: 0.5rem 0;
+    display: flex;
+    gap: 0.5rem;
+    flex-direction: column;
+  }
+}
