+++ carbon/components/Layout/DesktopSidebar/styles.module.scss
@@ -1,0 +1,80 @@
+@use "~@klimadao/lib/theme/breakpoints.scss";
+@use "theme/typography.module.scss";
+
+.desktopSidebar {
+  display: none;
+
+  @include breakpoints.desktop {
+    position: sticky;
+    top: 1rem;
+    padding: 4.8rem 2rem;
+    gap: 4rem;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--surface-04);
+  }
+
+  [aria-describedby="title"] {
+    text-align: center;
+    padding: 2rem 0;
+    border-top: 0.1rem solid var(--lightmode-font-01);
+    border-bottom: 0.1rem solid var(--lightmode-font-01);
+    font-family: Inter;
+    font-weight: 500;
+    font-size: 1.6rem;
+    line-height: 2rem;
+  }
+
+  [aria-describedby="links"] {
+    font-size: 1.4rem;
+    display: grid;
+    gap: 1.6rem;
+    padding-bottom: 4rem;
+    border-bottom: 0.1rem solid var(--lightmode-font-01);
+    white-space: nowrap;
+  }
+}
+
+.desktopSidebarItem {
+  min-width: 18rem;
+  height: 4.8rem;
+  padding: 0 0.8rem;
+  display: flex;
+  align-items: center;
+  text-align: left;
+  text-decoration: none;
+  gap: 1.2rem;
+  color: #767676;
+
+  &[aria-selected="true"] {
+    background-color: var(--lightmode-surface-02);
+    border-radius: 0.8rem;
+    color: var(--lightmode-font-01);
+
+    /**
+     * @todo this is kinda hacky */
+    span {
+      background-color: var(--brand-blue);
+      color: var(--brand-white);
+    }
+  }
+}
+
+.desktopSidebarBrand {
+  display: grid;
+  place-content: center;
+}
+
+.desktopSidebarIcon {
+  border-radius: 0.6rem;
+  background-color: var(--lightmode-surface-02);
+  color: var(--lightmode-font-01);
+  width: 3.2rem;
+  height: 3.2rem;
+
+  svg {
+    width: 2rem;
+    height: 2rem;
+    margin: 0.6rem;
+  }
+}
