+++ carbon/components/Layout/MobileBottomNav/index.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { navItems } from "./NavItems";
+import { navItems } from "../NavItems";
 import styles from "./styles.module.scss";
 
 import { MobileBottomNavItem } from "./MobileBottomNavItem";
