+++ carbon/components/Layout/MobileHeader/index.tsx
@@ -4,7 +4,7 @@ import styles from "./styles.module.scss";
 import { CarbonmarkDataLogo } from "components/Graphics/CarbonmarkDataLogo";
 import { ChangeLanguageButton } from "components/Layout/ChangeLanguageButton";
 import Link from "components/Link";
-import { MobileMenuButton } from "./MobileMenuButton";
+import { MobileMenuButton } from "../MobileMenuButton";
 
 export const MobileHeader: FC = () => {
   return (
