+++ carbon/components/Layout/styles.module.scss
@@ -1,179 +1,6 @@
 @use "~@klimadao/lib/theme/breakpoints.scss";
 @use "theme/typography.module.scss";
 
-.desktopSidebar {
-  display: none;
-
-  @include breakpoints.desktop {
-    position: sticky;
-    top: 1rem;
-    padding: 4.8rem 2rem;
-    gap: 4rem;
-    display: flex;
-    flex-direction: column;
-    background-color: var(--surface-04);
-  }
-
-  [aria-describedby="title"] {
-    text-align: center;
-    padding: 2rem 0;
-    border-top: 0.1rem solid var(--lightmode-font-01);
-    border-bottom: 0.1rem solid var(--lightmode-font-01);
-    font-family: Inter;
-    font-weight: 500;
-    font-size: 1.6rem;
-    line-height: 2rem;
-  }
-
-  [aria-describedby="links"] {
-    font-size: 1.4rem;
-    display: grid;
-    gap: 1.6rem;
-    padding-bottom: 4rem;
-    border-bottom: 0.1rem solid var(--lightmode-font-01);
-    white-space: nowrap;
-  }
-}
-
-.desktopSidebarItem {
-  min-width: 18rem;
-  height: 4.8rem;
-  padding: 0 0.8rem;
-  display: flex;
-  align-items: center;
-  text-align: left;
-  text-decoration: none;
-  gap: 1.2rem;
-  color: #767676;
-
-  &[aria-selected="true"] {
-    background-color: var(--lightmode-surface-02);
-    border-radius: 0.8rem;
-    color: var(--lightmode-font-01);
-
-    /**
-     * @todo this is kinda hacky */
-    span {
-      background-color: var(--brand-blue);
-      color: var(--brand-white);
-    }
-  }
-}
-
-.desktopSidebarBrand {
-  display: grid;
-  place-content: center;
-}
-
-.desktopSidebarIcon {
-  border-radius: 0.6rem;
-  background-color: var(--lightmode-surface-02);
-  color: var(--lightmode-font-01);
-  width: 3.2rem;
-  height: 3.2rem;
-
-  svg {
-    width: 2rem;
-    height: 2rem;
-    margin: 0.6rem;
-  }
-}
-
-.topBarButton {
-  min-width: 4.8rem;
-  min-height: 4.8rem;
-  width: 4.8rem;
-  height: 4.8rem;
-  color: var(--lightmode-font-01);
-  background-color: var(--lightmode-surface-02);
-
-  svg {
-    width: 2.4rem;
-    height: 2.4rem;
-  }
-}
-
-.topBarTooltip {
-  @include typography.body1;
-  text-align: left;
-  background: var(--brand-white);
-  border-radius: 0.8rem;
-  padding: 0 1.2rem;
-  min-width: 15rem;
-  box-shadow:
-    0 0.4rem 0.4rem rgba(0, 0, 0, 0.24),
-    0 0 0.4rem rgba(0, 0, 0, 0.12);
-
-  @media (max-width: 32.5rem) {
-    &.tippy-box {
-      max-width: 24rem !important; // fix x-overflow bug on small mobile
-    }
-  }
-
-  div[aria-describedby="tooltip-content"] {
-    padding: 0.5rem 0;
-    display: flex;
-    gap: 0.5rem;
-    flex-direction: column;
-  }
-}
-
-.mobileHeader {
-  height: 6.4rem;
-  display: flex;
-  align-items: center;
-  padding: 0.8rem 1.6rem 0.8rem 1.6rem;
-  justify-content: space-between;
-
-  @include breakpoints.desktop {
-    display: none;
-  }
-}
-
-.mobileNavButtons {
-  display: flex;
-  gap: 1.6rem;
-}
-
-.mobileBottomNav {
-  position: sticky;
-  bottom: 0;
-  z-index: 700;
-  display: flex;
-  width: 100%;
-  background-color: var(--darkmode-surface-02);
-  flex-direction: row;
-  justify-content: space-evenly;
-
-  @include breakpoints.desktop {
-    display: none;
-  }
-
-  > * {
-    flex: 1;
-  }
-}
-
-.mobileBottomNavItem {
-  display: flex;
-  padding: 0.4rem;
-  height: 3.6rem;
-  justify-content: center;
-  align-items: center;
-
-  &[aria-selected="true"] {
-    background-color: var(--darkmode-surface-03);
-    color: var(--brand-blue);
-  }
-
-  color: var(--darkmode-font-03);
-
-  svg {
-    width: 2rem;
-    height: 2rem;
-  }
-}
-
 .main {
   padding: 0 1.6rem;
   flex: 1;
@@ -190,64 +17,25 @@
   background-color: var(--brand-background);
 }
 
-.footer {
-  background-color: var(--darkmode-surface-01);
-  color: var(--brand-white);
-  padding: 4rem 2.4rem;
-
-  display: grid;
-  gap: 4rem;
-  align-items: center;
-
-  @include breakpoints.desktopLarge {
-    display: flex;
-    justify-content: space-between;
-  }
-
-  a {
-    text-decoration: none;
-    color: var(--brand-white);
-  }
-}
-
-.footerNavLinks {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  gap: 1.6rem;
-
-  @include breakpoints.desktop {
-    gap: 3.2rem;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-
-  a {
-    white-space: nowrap;
-  }
-}
-
-.footerSocialLinks {
-  display: flex;
-  gap: 1.6rem;
-  align-items: center;
-  justify-content: space-between;
-
-  @include breakpoints.large {
-    justify-content: center;
-  }
-
-  svg {
-    width: 1.6rem;
-    height: 1.6rem;
-  }
-}
-
 .desktopLayout {
   @include breakpoints.desktop {
     align-items: start;
     display: grid;
     grid-template-columns: min-content auto;
     min-height: 100vh;
+  }
+}
+
+.topBarButton {
+  min-width: 4.8rem;
+  min-height: 4.8rem;
+  width: 4.8rem;
+  height: 4.8rem;
+  color: var(--lightmode-font-01);
+  background-color: var(--lightmode-surface-02);
+
+  svg {
+    width: 2.4rem;
+    height: 2.4rem;
   }
 }
